Move the VHS into a separate storage package
