+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/transformer/utils/TransformableMessageUtils.scala
@@ -15,7 +15,7 @@ import uk.ac.wellcome.models.transformable.sierra.{
   SierraBibRecord,
   SierraItemRecord
 }
-import uk.ac.wellcome.storage.HybridRecord
+import uk.ac.wellcome.storage.vhs.HybridRecord
 import uk.ac.wellcome.utils.JsonUtil
 import uk.ac.wellcome.test.fixtures.S3.Bucket
