+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/vhs/HybridRecordEnricher.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.storage
+package uk.ac.wellcome.storage.vhs
 
 import shapeless.ops.hlist.{Collect, Prepend, Zip}
 import shapeless.{Id => ShapelessId, _}
